addding the modules
